chatbot/requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
langchain-google-genai
google-generativeai
python-dotenv   
langchain
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
pypdf
PyPDF2
pydantic
requests
tiktoken
python-docx   # (optional for docx support)
python-multipart